updated to typescript
