+++ index.html
@@ -1,7 +1,6 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <script src="/phaser.min.js"></script>
     <script src="/dist/main.js"></script>
 </head>
 <body>
