+++ package.json
@@ -6,11 +6,15 @@
   "dependencies": {},
   "devDependencies": {
     "http-server": "^0.12.3",
+    "phaser": "^3.24.1",
+    "ts-loader": "^8.0.6",
+    "typescript": "^4.0.3",
     "webpack": "^5.1.3",
     "webpack-cli": "^4.1.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "watch": "npx webpack --watch & npx http-server",
     "serve": "npx http-server",
     "build": "npx webpack"
   },
