+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "target": "es5"
+  },
+  "include": [
+    "src/*"
+  ]
+}
