+++ webpack.config.js
@@ -1,9 +1,21 @@
 const path = require('path');
-
 module.exports = {
-  entry: './src/index.js',
+  entry: './src/index.ts',
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/
+      }
+    ]
+  },
+  resolve: {
+    extensions: [ '.ts', '.tsx', '.js' ]
+  },
   output: {
     filename: 'main.js',
-    path: path.resolve(__dirname, 'dist'),
+    path: path.resolve(__dirname, 'dist')
   },
+  mode: 'development'
 };
